fix(web): satisfy workspace editor CI gates
Summary:
- split the Project AI workspace browser into focused editor-state and pane
  modules so the new editor stays within frontend file-budget limits
- document the validated local workspace read paths for gosec and keep the
  backend save flow lint-clean
- register the new workspace file save route as an intentional CLI/OpenAPI
  parity gap until a first-class CLI command exists

Rationale:
- the feature branch was functionally complete, but the CI-aligned publish gate
  still failed on file-budget, gosec, and CLI parity checks
- extracting the editor logic keeps the shipped behavior unchanged while making
  the implementation easier to maintain and publish safely

Tests:
- PATH=$HOME/.nvm/versions/node/v22.22.1/bin:$PATH pnpm --dir web check
- PATH=$HOME/.nvm/versions/node/v22.22.1/bin:$PATH pnpm --dir web exec vitest run src/lib/features/chat/project-conversation-workspace-browser-navigation.test.ts src/lib/features/chat/project-conversation-workspace-browser-refresh.test.ts src/lib/features/chat/project-conversation-workspace-browser-state.test.ts
- OPENASE_GO_TEST_PROGRESS_MODE=json PATH=$HOME/.local/go1.26.1/bin:$PATH make check
- PLAYWRIGHT_WEB_PORT=4273 PLAYWRIGHT_PORT=4273 PATH=$HOME/.nvm/versions/node/v22.22.1/bin:$PATH pnpm --dir web exec playwright test --reporter=dot --quiet
- PATH=$HOME/.local/go1.26.1/bin:$PATH make build
- PATH=$HOME/.local/go1.26.1/bin:$PATH make LINT_BASE_REV=f202584bf85c1af29282c8d3be58604fcb9ee6f4 lint
- PATH=$HOME/.local/go1.26.1/bin:$PATH make lint-depguard
- make lint-architecture
