+++ web/src/lib/features/chat/project-conversation-workspace-browser.svelte
@@ -1,6 +1,5 @@
 <script lang="ts">
-  /* eslint-disable max-lines */
-  import { untrack } from 'svelte'
+  import { onDestroy, untrack } from 'svelte'
   import { Button } from '$ui/button'
   import { cn } from '$lib/utils'
   import { appStore } from '$lib/stores/app.svelte'
@@ -15,12 +14,9 @@
   } from '@lucide/svelte'
   import type { ProjectConversationWorkspaceDiff } from '$lib/api/chat'
   import { PROJECT_AI_FOCUS_PRIORITY } from './project-ai-focus'
-  import WorkspaceTerminalPanel from './workspace-terminal-panel.svelte'
-  import ProjectConversationWorkspaceBrowserDetail from './project-conversation-workspace-browser-detail.svelte'
-  import ProjectConversationWorkspaceBrowserSidebar from './project-conversation-workspace-browser-sidebar.svelte'
+  import ProjectConversationWorkspaceBrowserPane from './project-conversation-workspace-browser-pane.svelte'
   import { createProjectConversationWorkspaceBrowserState } from './project-conversation-workspace-browser-state.svelte'
   import { createTerminalManager } from './terminal-manager.svelte'
-  import { onDestroy } from 'svelte'
 
   let {
     conversationId = '',
@@ -69,67 +65,6 @@
     navigator.clipboard.writeText(path)
     pathCopied = true
     setTimeout(() => (pathCopied = false), 1500)
-  }
-
-  // -- Sidebar resize --
-  const MIN_SIDEBAR_WIDTH = 180
-  const MAX_SIDEBAR_WIDTH = 480
-  let sidebarWidth = $state(240)
-  let sidebarResizing = $state(false)
-
-  function handleSidebarResizeStart(event: PointerEvent) {
-    event.preventDefault()
-    sidebarResizing = true
-    const startX = event.clientX
-    const startWidth = sidebarWidth
-
-    function onMove(e: PointerEvent) {
-      sidebarWidth = Math.min(
-        MAX_SIDEBAR_WIDTH,
-        Math.max(MIN_SIDEBAR_WIDTH, startWidth + (e.clientX - startX)),
-      )
-    }
-
-    function onUp() {
-      sidebarResizing = false
-      window.removeEventListener('pointermove', onMove)
-      window.removeEventListener('pointerup', onUp)
-    }
-
-    window.addEventListener('pointermove', onMove)
-    window.addEventListener('pointerup', onUp)
-  }
-
-  // -- Terminal panel vertical resize --
-  const MIN_TERMINAL_HEIGHT = 120
-  const DEFAULT_TERMINAL_HEIGHT = 260
-  let terminalHeight = $state(DEFAULT_TERMINAL_HEIGHT)
-  let terminalResizing = $state(false)
-  let containerElement: HTMLDivElement | null = null
-
-  function handleTerminalResizeStart(event: PointerEvent) {
-    event.preventDefault()
-    terminalResizing = true
-    const startY = event.clientY
-    const startHeight = terminalHeight
-
-    function onMove(e: PointerEvent) {
-      const maxHeight = containerElement ? containerElement.clientHeight - 100 : 600
-      terminalHeight = Math.min(
-        maxHeight,
-        Math.max(MIN_TERMINAL_HEIGHT, startHeight - (e.clientY - startY)),
-      )
-    }
-
-    function onUp() {
-      terminalResizing = false
-      window.removeEventListener('pointermove', onMove)
-      window.removeEventListener('pointerup', onUp)
-      terminalManager.refitAll()
-    }
-
-    window.addEventListener('pointermove', onMove)
-    window.addEventListener('pointerup', onUp)
   }
 
   let refreshGeneration = $state(0)
@@ -262,7 +197,6 @@
 <div
   class="bg-background flex h-full min-h-0 w-full flex-col"
   data-testid="project-conversation-workspace-browser"
-  bind:this={containerElement}
 >
   <!-- Compact toolbar -->
   <div class="border-border flex h-9 items-center gap-1.5 border-b px-3">
@@ -350,85 +284,12 @@
       The workspace will appear after Project AI provisions the conversation workdir.
     </div>
   {:else}
-    <div
-      class={cn(
-        'flex min-h-0 flex-1 flex-col',
-        (sidebarResizing || terminalResizing) && 'select-none',
-      )}
-    >
-      <!-- Files area -->
-      <div class="flex min-h-0 flex-1">
-        <!-- Sidebar (resizable) -->
-        <div
-          class="relative min-h-0 shrink-0 overflow-hidden"
-          style="width: {sidebarWidth}px"
-          data-testid="workspace-browser-sidebar-panel"
-        >
-          <ProjectConversationWorkspaceBrowserSidebar
-            repos={browser.metadata?.repos ?? []}
-            selectedRepoPath={browser.selectedRepoPath}
-            {selectedRepo}
-            {selectedRepoDiff}
-            treeNodes={browser.treeNodes}
-            expandedDirs={browser.expandedDirs}
-            loadingDirs={browser.loadingDirs}
-            selectedFilePath={browser.selectedFilePath}
-            onOpenRepo={browser.openRepo}
-            onToggleDir={browser.toggleDir}
-            onSelectFile={browser.selectFile}
-          />
-          <!-- Resize handle -->
-          <div
-            class={cn(
-              'absolute inset-y-0 right-0 z-10 w-1 cursor-col-resize transition-colors',
-              sidebarResizing ? 'bg-primary' : 'bg-border hover:bg-primary/50',
-            )}
-            role="separator"
-            aria-orientation="vertical"
-            onpointerdown={handleSidebarResizeStart}
-          ></div>
-        </div>
-        <!-- Detail -->
-        <div
-          class="min-h-0 min-w-0 flex-1 overflow-hidden"
-          data-testid="workspace-browser-detail-panel"
-        >
-          <ProjectConversationWorkspaceBrowserDetail
-            {selectedRepo}
-            selectedFilePath={browser.selectedFilePath}
-            preview={browser.preview}
-            patch={browser.patch}
-            editorState={browser.selectedEditorState}
-            draftDiff={browser.selectedDraftDiff}
-            fileLoading={browser.fileLoading}
-            fileError={browser.fileError}
-            {runtimeActive}
-            onViewModeChange={browser.setSelectedViewMode}
-            onDraftChange={browser.updateSelectedDraft}
-            onSave={() => void browser.saveSelectedFile()}
-            onRevert={browser.revertSelectedDraft}
-            onKeepDraft={browser.keepSelectedDraft}
-            onReloadSavedVersion={browser.reloadSelectedSavedVersion}
-          />
-        </div>
-      </div>
-
-      <!-- Terminal panel (bottom, like VSCode) -->
-      {#if terminalManager.panelOpen}
-        <!-- Resize handle -->
-        <div
-          class={cn(
-            'h-[3px] shrink-0 cursor-row-resize transition-colors',
-            terminalResizing ? 'bg-primary' : 'bg-border hover:bg-primary/50',
-          )}
-          role="separator"
-          aria-orientation="horizontal"
-          onpointerdown={handleTerminalResizeStart}
-        ></div>
-        <div class="flex min-h-0 shrink-0 overflow-hidden" style="height: {terminalHeight}px">
-          <WorkspaceTerminalPanel manager={terminalManager} />
-        </div>
-      {/if}
-    </div>
+    <ProjectConversationWorkspaceBrowserPane
+      {browser}
+      {selectedRepo}
+      {selectedRepoDiff}
+      {runtimeActive}
+      {terminalManager}
+    />
   {/if}
 </div>
